Experimental classifier
The classifier mode can be switched via command line argument -C
First one is very close to original classifier (which seems to be a variant of DBRS
algorithm for for density based clustering).
Second and third are the same but merging of clusters is continuing till nothing
can be merged anymore. That dramatically increase CPU usage so some memory is
allocated to cache operations. So both CPU usage and RAM usage are much higher
in this mode. On the other hand this increase a lot quality of symbols clusterization
and thus allows to produce up to 30% smaller djvu files.